Coverage (#34)
* Enable coverage in cmake

Adds the cmake targets `coverage` and `coverage_html`. The first
generates an xml report and is called by the github action to upload
coverage data to codecov and coverity. The former can be used to
generate a coverage report locally outside of the CI.

Coverage is only evaluated on linux for now.

Additionally gcc in the CI is upgraded from version 9 to 11.

* Enable coverage by default and exclude concepts