majordomo tests: Make heartbeat test more robust
Be more lenient with the asserted heartbeat interval, as the broker
can go into polling just before it would normally send the heartbeats,
then it can take longer than two heartbeat intervals for the worker to
receive the heartbeat. Also, all in all the number of heartbeats sent
by the broker before dropping an unresponsive worker depends on the
liveness, so take that into account.