added mmi.service & made worker unit-tests more robust
* Implement mmi.service

Implement mmi.service: It can be queried whether a certain service name
is known (i.e. a worker registered), and also list all known services.

Use it to simplify tests, waiting on startup until the worker is
registered, instead of handling the possible "unknown service" replies.


* majordomo tests: Timeout reading replies after 3 seconds
* majordomo tests: Make heartbeat test more robust

Be more lenient with the asserted heartbeat interval, as the broker
can go into polling just before it would normally send the heartbeats,
then it can take longer than two heartbeat intervals for the worker to
receive the heartbeat. Also, all in all the number of heartbeats sent
by the broker before dropping an unresponsive worker depends on the
liveness, so take that into account.

* Add test for BasicMdpWorker sending heartbeat and reconnect

* BasicMdpWorker: send DISCONNECT on shutdown

send disconnect to broker when shutting down the worker.

* Tests: Use tryReadOne directly

The only way to get a useful error message and position when a message
isn't received seems to REQUIRE the optional message to have a value
on every usage.