MessageFrame: fix overzealos assertion
It is a valid state for zeromq send operations to return EAGAIN(11)
error codes that the user has to handle by resending later.
This is also handled correctly in higher level code, but in Debug mode
this didn't work because the EAGAIN error would trigger the assertion.

Also incorporate clang-tidy feedback.