BasicThreadPool example: adjust to new libfmt
had to use fmt::runtime here even though everything should be known at
compile time... the combination of fmt and the template logic in
ThreadPool::execute makes this difficult to grasp. I'm a bit suspicious
of the FWD macro but not sure how it would have to be changed.
But since it is only a sample and runtime format strings were the
default before I think this should be acceptable.