RestBackend: Use real zmq subscription
The previous implementation used the Request Socket to send a
subscription mdp command instead of using the broker's pub/sub endpoint.
Also the naming and usage of the different Sockets was inconsistent.

This changes the socket types and uses the SUB socket for subscriptions
while using the dealer socket for get requests.