Avoid cherry-picking with different user
Some people configure git to GPG-sign each commit by default.
For these users forcing the cherry-pick with a different anonymous user
would fail, as they wouldn't have a GPG key generated for Anonymous <>.

To fix this avoid committing the cherry-pick and instead just apply the
diff (just like patch(1) would).
This also simplifies the cmake code as we no longer need to modify the
user (that code existed to fix the build for people that didn't
configure a git user yet).