Handle https subscriptions, fix cert store bug
https subscriptions were not handled, create a SSLCLient and reuse
existing code for subscribing.

the certificate store was shared between all SslClients in our
RestClient, this does not work, because the client owns the store and
deletes it on its own destruction.

this patch also adds the certificate specified by OPENCMW_REST_CERT_FILE
to our RestDefaultCLientCertificates to make it trusted.