RestBackend: Do not expire connections that are in use
Actually update the "lastUsed" timestamp when accessing a cached
connection. Otherwise the cache is destroyed and recreated in normal
operation, leading to LongPollingIdx=Next requests to fail (as they
were redirected to the latest index before cache reset, which is then
considered a "future" index, as the latest index is reset to 0).