RestClient: Correctly use LongPollingIdx Parameter
Changes the RestClient to internally track the last received update
index instead of just blindly always requesting the latest update.
This prevents duplicate notifications and allows catching up and
blocking on not yet arrived updates.

For emscripten this needs a bit of trickery because the api of
emscripten_fetch does not provide a way to access the final redirected
URL, so we have to call out to javascript to retrieve it.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>