RestBackend: Remove artificial rate limit
The poll loop that handles the messages from the broker had a 100ms
sleep built-in and only handled only one update per iteration and
connection.
This lead to updates being acumulated in the RestBackend's subscription
socket, whenever the underlying property would update with more than
10Hz.

This commit reduces the sleep between poll loops to 10ms and handles all
notifications for all connections at once, allowing the RestBackend to
catch up on a burst of notifications.

smaller changes:
- make http headers case-insensitive
- general code-cleanup

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>