Implement RBAC integration
This combines RBAC with OAuth authorization, so that general mdp
messages can be signed in a cryptographically secure way, such that any
arbitrary receiver (such as a Worker) can verify that the message
belongs to a user that has access to a certain role.

This is implemented with state of the art elliptic curve cryptography by
signing the message with a private key, and uploading the public key to
a key store worker which hands out all public keys as a service via
access over the key's hash.
Any receiver can then retrieve the public key by just asking for the
key's hash, which is always included in the RBAC data field of any
signed message (plus of course the data needed for signing).

Special care is taken to avoid cross-referencing public keys to specific
users. As soon as the OAuth flow is finished, any public key can no
longer be used to pinpoint a specific user, instead the keystore only
holds the data which role this key belongs to.

This implements and closes #349.

Signed-off-by: Magnus Groß <magnus.gross@kdab.com>