EmscriptenRestClient: proxy fetch to main thread
Fetch calls on pthreads will only be handled once the thread terminates
and the control of the underlying web-worker is given back to the
emscripten runtime, which most of the time never happens as they belong
to a static thread-pool.

The fact that most of the subscriptions worked before was because the
start handler or settings_updated handler was not called by the block
worker, but by the main thread setting up the graph.

For some reason it was necessary to increase the stack size for this
proxying to work.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>