RestBackend: redirect requests for old longpolling index to newest
Changes the behaviour for requests for old long-polling indices that are
no longer cached from returning an error to returning a redirect to the
latest sample.
This has to be followed up on the client side to produce a
warning/marker about the incontinuity whenever a request is forwarded to
a different index than requested.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>