CMake: fix CMakeWarnings and link statically against zeromq
Fixes some easy to fix CMake Warnings to make the cmake generation less
spammy.

Also I noticed that up to now we were linking against a custom build of
zeromq dynamically which would either require to also copy the
libzeromq.so dynamic library or rely on the library on the target system
to be binary identical to our internal build.
This changes the linkage to static so we consistently link dynamically
to system libraries and statically to the libraries we build as part of
our build (openssl, sodium and zeromq at the moment).

Also we directly use and build libsodium, but zeromq was replacing it
with libtweetnacl, which conflicted at link time to this commit makes
our build of libsodium usable by zeromq by exporting the appropriate
variables.
Since libsodium is not compiled as fPIC, the shared library build of
zeromq had to be disabled as trying to link libzmq.so against the non
fPIC libsodium caused linker errors.

Signed-off-by: Alexander Krimm <alex@wirew0rm.de>