RestClientEmscripten: allow long-polling catch-up redirect
Allow the RestClient to catch-up if the service sends a redirect
response to the current long polling index.
Before the client would cause a connection error if the index did not
match the expectation. The result is still propagated to the user to be
able to show a visual indication of the problem.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>