Terminate pub/sub subscription URIs with "#"
In the ColorWorker example ("Subscriptions" test in
majordomoworker_rest_tests), the broker with (X)PUB client subscriptions
"/colors", "/colors?green" and "/colors?red" has to forward a message
with topic "/colors?green&red" to "/colors?green" and "/colors&red", to
ensure that all clients receive the message. Due to the subscription
logic of PUB/SUB sockets the client subscribing to "/colors" would
retrieve the message twice though, and couldn't identify the
duplicates.

To prevent this, append "#" to the ZMQ-level subscription strings,
("/colors#", "/colors?green#", "/colors?red#).

Signed-off-by: Frank Osterfeld <frank.osterfeld@kdab.com>