Clang: Enable libc++ and require clang-18
Ensure libc++ is used when using clang, in preparation for using e.g.
std::expected. Bump minimum version to clang 18.

Make sure mdp::Topic::hash() is equal for equal instances, the
implementation of hash() iterating the params in an undefined order
assumed hash_combine/std::hash to be commutative, which doesn't seem
to be the case when using libc++'s std::hash.

Signed-off-by: Frank Osterfeld <frank.osterfeld@kdab.com>